[Refactor] 클래스명 일치하도록 수정
