+++ src/main/java/shop/bookbom/front/domain/category/adapter/CategoryAdapter.java
@@ -1,4 +1,4 @@
-package shop.bookbom.front.domain.category.adaptor;
+package shop.bookbom.front.domain.category.adapter;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -10,7 +10,7 @@ import shop.bookbom.front.domain.category.dto.response.CategoryDepthResponse;
 import shop.bookbom.front.domain.category.dto.response.CategoryNameAndChildResponse;
 
 @FeignClient(value = "BOOKBOM-FRONT-CATEGORY", path = "/shop", url = "${bookbom.gateway-url}")
-public interface CategoryAdaptor {
+public interface CategoryAdapter {
 
     @GetMapping("/category/all")
     CommonResponse<CategoryDepthResponse> getAllCategories();
